# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.Response:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  models.Timestamp:
    properties:
      time.Time:
        type: string
    type: object
  models.Transaction:
    properties:
      account_id:
        type: integer
      account_name:
        description: Computed fields
        type: string
      allocated:
        type: integer
      amount:
        type: integer
      contact_id:
        type: integer
      contact_name:
        type: string
      created_at:
        $ref: '#/definitions/models.Timestamp'
      description:
        type: string
      id:
        type: integer
      reference:
        type: string
      transaction_date:
        type: string
      transfer_account_id:
        type: integer
      transfer_account_name:
        type: string
      type:
        description: income, expense, transfer
        type: string
      unallocated:
        type: integer
      updated_at:
        $ref: '#/definitions/models.Timestamp'
    type: object
  models.TransactionDocument:
    properties:
      amount:
        type: integer
      created_at:
        $ref: '#/definitions/models.Timestamp'
      document_id:
        type: integer
      document_type:
        description: bill, invoice, payout or recurring_payment_occurrence
        type: string
      id:
        type: integer
      transaction_id:
        type: integer
    type: object
  models.TransactionDocumentInput:
    properties:
      amount:
        type: integer
      document_id:
        type: integer
      document_type:
        type: string
    type: object
  models.TransactionInput:
    properties:
      account_id:
        type: integer
      amount:
        type: integer
      contact_id:
        type: integer
      description:
        type: string
      reference:
        type: string
      transaction_date:
        type: string
      transfer_account_id:
        type: integer
      type:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: API for managing accounts, contacts, bills, invoices, and transactions.
  title: Portal API
  version: 1.0.0
paths:
...
  /transactions:
    get:
      description: Get a list of all bank transactions (income, expense, transfer)
        with allocation info.
      parameters:
      - description: Filter by account
        in: query
        name: account_id
        type: integer
      - description: Filter by contact
        in: query
        name: contact_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Transaction'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Create a new bank transaction (income, expense, or transfer).
      parameters:
      - description: Transaction contents
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.TransactionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Transaction'
              type: object
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - transactions
  /transactions/{id}:
    delete:
      description: Remove a transaction.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete transaction
      tags:
      - transactions
    get:
      description: Get details and allocation status of a specific transaction.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Transaction'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get transaction
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: Update details of an existing transaction.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated transaction contents
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.TransactionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Transaction'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update transaction
      tags:
      - transactions
...
  /transactions/{id}/links:
    get:
      description: Get all bills and invoices linked (paid) by this transaction.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.TransactionDocument'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List transaction links
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Allocate an amount from a transaction to a specific bill or invoice.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Link details
        in: body
        name: link
        required: true
        schema:
          $ref: '#/definitions/models.TransactionDocumentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.TransactionDocument'
              type: object
      security:
      - BearerAuth: []
      summary: Create transaction link
      tags:
      - transactions
  /transactions/{id}/links/{linkId}:
    delete:
      description: Deallocate an amount from a transaction to a bill or invoice.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Link ID
        in: path
        name: linkId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Delete transaction link
      tags:
      - transactions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"